add users_list.js && refactor

+++ src/utils/fs_assistant.js
@@ -25,6 +25,7 @@ export const writeFile = (path, data) => {
 };
 
 export const deleteFile = path => {
+  if (!existItem(path)) return null;
   fs.unlinkSync(path);
 };
